kernel/pyproject.toml
[tool.poetry]
name = "kernel"
version = "0.1.0"
description = ""
authors = ["Harry <harry@myntlabs.io>"]
readme = "README.md"
packages = [
    { include = "kernel" }
]

[tool.poetry.dependencies]
python = "^3.12"
jupyterlab = "^4.4.0"
ipykernel = "^6.29.5"
boto3 = ">=1.0.0,<1.36.0"
s3contents = "^0.11.2"
polars = "^1.27.1"
deltalake = "^0.25.5"
ipydatagrid = "^1.4.0"
jupyterlab-widgets = "^3.0.14"
ipywidgets = "^8.1.6"
jupyterlab-lsp = "^5.1.0"
python-lsp-server = { version = "^1.12.2", extras = ["all"] }
python-lsp-ruff = "^2.2.2"
perspective-python = "^3.5.1"
fsspec = "^2025.3.2"
s3fs = "^2025.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
builtins = ["write_delta", "read_delta", "scan_delta", "drop_delta", "display", "new_run", "WORKSPACE_NAME"]